refactor(@binstruct/arp,@binstruct/icmp,@binstruct/inet,@binstruct/ipv4,@binstruct/udp): centralize stack composition in @binstruct/inet
Each protocol package now decodes only its own layer — header, fields,
trailers, payload as Uint8Array — with no awareness of what carries it.
@binstruct/inet is the only package that knows how to compose layers
into a stack.

Removed from protocol packages:
- @binstruct/ipv4: ipv4Refiner factory + FrameWithIpv4 type, Ethernet2Frame import
- @binstruct/udp: udpRefiner + Ipv4WithUdp type, Ipv4Datagram import
- @binstruct/icmp: icmpRefiner + Ipv4WithIcmp type, refiners.ts file, Ipv4Datagram import
- @binstruct/arp: arpRefiner + FrameWithArp type, Ethernet2Frame import

Added to @binstruct/inet:
- udpRefiner, icmpRefiner: Ipv4Datagram → typed L4 payload
- arpRefiner: Ethernet2Frame → typed ARP payload
- ipv4Refiner(coder): Ethernet2Frame → typed IPv4 payload (still takes
  the inner coder so callers can use ipv4Datagram directly or the
  fully-refined ipv4Coder)

Each protocol package now imports only from @hertzg/binstruct (and its
own utility deps like @hertzg/ip / @hertzg/mac). No cross-protocol
dependencies — udp doesn't know about IPv4, IPv4 doesn't know about
Ethernet, etc. The downward-only composition story is preserved
strictly within @binstruct/inet.
